feat: move common .js files to .ts
