--- src/common/httpMethods.js
@@ -1,8 +1,0 @@
-const HTTP_METHODS = {
-    GET: 'GET',
-    POST: 'POST',
-    PUT: 'PUT',
-    DELETE: 'DELETE',
-};
-
-module.exports = HTTP_METHODS;
