+++ src/common/httpMethods.ts
@@ -1,0 +1,8 @@
+const HTTP_METHODS = {
+  GET: 'GET',
+  POST: 'POST',
+  PUT: 'PUT',
+  DELETE: 'DELETE',
+};
+
+export default HTTP_METHODS;
